Rollback create-react-app from v1.0 to v0.9
create-react-app v1.0 force the import path  to be inside project root,
which prevents us from importing redux-draft lib into the doc site. A
quick solution is to use `npm link` to create symbolic link, but
unfortunately this somehow breaks the build process of redux-draft lib.
The current proper way is to rollback the version of create-react-app.